1.1_5: 메인페이지 변경 및 fetch를 통한 통신 설정 완료 ** 0112 @Test annotation mismatch(jupyter.api x) ** ** 0114 BootStrap FrameWork ** ** 0115 login page proto // notice board 완성 ** ** 0116 회원가입 시퀀스 완성 ** ** 0123 메인 홈페이지 디자인 변경 ** ** 0125 메인 홈 get method 완료** ** 0126 관리자 페이지 생성 및 업로드 기능 완료 ** ** 0128 로그인 시스템 및 관리자 권한 완료 ** ** 0128-2 공지사항 등록 및 item 로직 변경(팩토리 메소드 패턴 적용) **
** 0129 파일저장 시스템 변경 및 api 리팩토링 + 이미지 파일 외부로 전환(메인 디렉토리 하위 img)+공지사항 조회기능 추가 및 post url 변경**

**0131 FAQ조회 및 삭제 기능 완성
  제품 조회,주문 관리조회,회원 관리 조회,배송 조회 만들어야함. **

+++ src/main/java/com/pipe09/OnlineShop/Repository/BoardRepository.java
@@ -3,13 +3,16 @@ package com.pipe09.OnlineShop.Repository;
 
 import com.pipe09.OnlineShop.Domain.Board.Notice;
 import lombok.RequiredArgsConstructor;
+import org.junit.Assert;
+import org.springframework.orm.jpa.JpaSystemException;
 import org.springframework.stereotype.Repository;
 
 import javax.persistence.EntityManager;
+import java.time.LocalDate;
 import java.util.List;
 
-@Repository
 @RequiredArgsConstructor
+@Repository
 public class BoardRepository {
     private final EntityManager em;
     public Long save(Notice notice){
@@ -19,9 +22,14 @@ public class BoardRepository {
     public List<Notice> readNoticeAll(){
         return em.createQuery("select nlist from Notice nlist").getResultList();
     }
-    public void removeNoticeByID(Long id){
+    public Notice findByID(Long id){ return em.find(Notice.class,id);}
+    public boolean removeNoticeByID(Long id) throws JpaSystemException {
         Notice notice=em.find(Notice.class,id);
+        if(notice==null){
+            return false;
+        }
         em.remove(notice);
-
+        return true;
     }
+
 }
